Pull receiver creation into component
