+++ src/NServiceBus.Core/StartableEndpoint.cs
@@ -4,7 +4,6 @@ namespace NServiceBus
     using System.Collections.Generic;
     using System.Security.Principal;
     using System.Threading.Tasks;
-    using ConsistencyGuarantees;
     using Features;
     using Logging;
     using ObjectBuilder;
@@ -39,7 +38,9 @@ namespace NServiceBus
             AppDomain.CurrentDomain.SetPrincipalPolicy(PrincipalPolicy.WindowsPrincipal);
 
             var mainPipeline = new Pipeline<ITransportReceiveContext>(builder, pipelineConfiguration.Modifications);
-            var receivers = CreateReceivers(mainPipeline);
+            var mainPipelineExecutor = new MainPipelineExecutor(builder, eventAggregator, pipelineCache, mainPipeline);
+
+            var receivers = receiving.CreateReceivers(mainPipelineExecutor, builder.Build<RecoverabilityExecutorFactory>(), eventAggregator, builder, criticalError);
             await InitializeReceivers(receivers).ConfigureAwait(false);
 
             var featureRunner = await StartFeatures(messageSession).ConfigureAwait(false);
@@ -92,72 +93,6 @@ namespace NServiceBus
             }
         }
 
-        List<TransportReceiver> CreateReceivers(IPipeline<ITransportReceiveContext> mainPipeline)
-        {
-            if (settings.GetOrDefault<bool>("Endpoint.SendOnly"))
-            {
-                return new List<TransportReceiver>();
-            }
-
-            var purgeOnStartup = settings.GetOrDefault<bool>("Transport.PurgeOnStartup");
-            if (purgeOnStartup)
-            {
-                Logger.Warn($"All queues owned by the '{settings.EndpointName()}' endpoint will be purged on startup.");
-            }
-
-            var errorQueue = settings.ErrorQueueAddress();
-            var requiredTransactionSupport = settings.GetRequiredTransactionModeForReceives();
-
-            var recoverabilityExecutorFactory = builder.Build<RecoverabilityExecutorFactory>();
-
-            var receivers = BuildMainReceivers(errorQueue, purgeOnStartup, requiredTransactionSupport, recoverabilityExecutorFactory, mainPipeline);
-
-            // ReSharper disable once LoopCanBeConvertedToQuery
-            foreach (var satellitePipeline in settings.Get<SatelliteDefinitions>().Definitions)
-            {
-                var satelliteRecoverabilityExecutor = recoverabilityExecutorFactory.Create(satellitePipeline.RecoverabilityPolicy, eventAggregator, satellitePipeline.ReceiveAddress);
-                var satellitePushSettings = new PushSettings(satellitePipeline.ReceiveAddress, errorQueue, purgeOnStartup, satellitePipeline.RequiredTransportTransactionMode);
-
-                receivers.Add(new TransportReceiver(satellitePipeline.Name, receiving.BuildMessagePump(), satellitePushSettings, satellitePipeline.RuntimeSettings, new SatellitePipelineExecutor(builder, satellitePipeline), satelliteRecoverabilityExecutor, criticalError));
-            }
-
-            return receivers;
-        }
-
-        List<TransportReceiver> BuildMainReceivers(string errorQueue, bool purgeOnStartup, TransportTransactionMode requiredTransactionSupport, RecoverabilityExecutorFactory recoverabilityExecutorFactory, IPipeline<ITransportReceiveContext> mainPipeline)
-        {
-            var recoverabilityExecutor = recoverabilityExecutorFactory.CreateDefault(eventAggregator, receiving.LocalAddress);
-            var pushSettings = new PushSettings(receiving.LocalAddress, errorQueue, purgeOnStartup, requiredTransactionSupport);
-            var mainPipelineExecutor = new MainPipelineExecutor(builder, eventAggregator, pipelineCache, mainPipeline);
-            var dequeueLimitations = GetDequeueLimitationsForReceivePipeline();
-
-            var receivers = new List<TransportReceiver>();
-            receivers.Add(new TransportReceiver(MainReceiverId, receiving.BuildMessagePump(), pushSettings, dequeueLimitations, mainPipelineExecutor, recoverabilityExecutor, criticalError));
-
-            if (receiving.InstanceSpecificQueue != null)
-            {
-                var instanceSpecificQueue = receiving.InstanceSpecificQueue;
-                var instanceSpecificRecoverabilityExecutor = recoverabilityExecutorFactory.CreateDefault(eventAggregator, instanceSpecificQueue);
-                var sharedReceiverPushSettings = new PushSettings(instanceSpecificQueue, errorQueue, purgeOnStartup, requiredTransactionSupport);
-
-                receivers.Add(new TransportReceiver(MainReceiverId, receiving.BuildMessagePump(), sharedReceiverPushSettings, dequeueLimitations, mainPipelineExecutor, instanceSpecificRecoverabilityExecutor, criticalError));
-            }
-
-            return receivers;
-        }
-
-        //note: this should be handled in a feature but we don't have a good
-        // extension point to plugin atm
-        PushRuntimeSettings GetDequeueLimitationsForReceivePipeline()
-        {
-            if (settings.TryGet(out MessageProcessingOptimizationExtensions.ConcurrencyLimit concurrencyLimit))
-            {
-                return new PushRuntimeSettings(concurrencyLimit.MaxValue);
-            }
-
-            return PushRuntimeSettings.Default;
-        }
-
         IMessageSession messageSession;
         IBuilder builder;
         FeatureActivator featureActivator;
@@ -171,7 +106,6 @@ namespace NServiceBus
         ReceiveComponent receiving;
         CriticalError criticalError;
 
-        const string MainReceiverId = "Main";
         static ILog Logger = LogManager.GetLogger<StartableEndpoint>();
     }
 }
